update proprietary makefiles - allow using local directory instead of adb to pull from - add qminvapi.so

+++ extract-files.sh
@@ -3,6 +3,11 @@
 VENDOR=xiaomi
 DEVICE=aries
 
+if [ $# -eq 1 ]; then
+    COPY_FROM=$1
+    test ! -d "$COPY_FROM" && echo error reading dir "$COPY_FROM" && exit 1
+fi
+
 BASE=../../../vendor/$VENDOR/$DEVICE/proprietary
 rm -rf $BASE/*
 
@@ -11,7 +16,11 @@ for FILE in `cat proprietary-blobs.txt | grep -v ^# | grep -v ^$`; do
     if [ ! -d $BASE/$DIR ]; then
         mkdir -p $BASE/$DIR
     fi
-    adb pull /system/$FILE $BASE/$FILE
+    if [ "$COPY_FROM" = "" ]; then
+      adb pull /system/$FILE $BASE/$FILE
+    else
+      cp $COPY_FROM/$FILE $BASE/$FILE
+    fi
 done
 
 ./setup-makefiles.sh
